refactor(cvaos): move the route solvers out of the data package
Issue: The data package held both the parsed CSV rows and about 1,200 lines
of BFS and Dijkstra route solving, so its name described only half of what
was in it. Generation never called that code, since regions.py builds its
regions from the same rows directly, yet it still shipped in the apworld.

Solution: The three solvers now live in tools/routing, which is documented as
developer tooling that may read from data but must never be read by it.
